Progress on Exercise 1

+++ exercises/ex01/numeric_operators.py
@@ -4,7 +4,7 @@ __author__ = "730245854"
 left: int = int(input("Left-hand side: "))
 right: int = int(input("Right-hand side: "))
 exponent: int = left ** right
-division: int = int(left / right) 
+division: float = left / right
 integer_division: int = left // right
 remainder_modulo: int = left % right
 print(str(left) + " ** " + str(right) + " is " + str(exponent))
